Move textbox related component into its own folder (#99)
* Move textbox related component into its own folder

* fix unit tests

* remove warning in TextBoxEdit.css
