+++ front-end/src/atoms/TextBox/TextBoxEdit.css
@@ -20,7 +20,7 @@
 
 .edit-header{
   display: flex;
-  justify-content: end;
+  justify-content: flex-end;
 }
 
 .delete-icon{
